server/requirements.txt
anyio==3.6.1
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
fastapi==0.81.0
h11==0.13.0
idna==3.3
pycparser==2.21
pydantic==1.10.1
PyJWT==2.4.0
python-dotenv==0.20.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
